Refactor of React Contexts/Providers
Assets balances were only fetched when all assets were fetched from the API. If the user received or spent assets in the between, new balance would not be reflected on the app.

To fix this, balances were moved to a React Context, and are updated every time the app receives a NEW_TX event from Marina. This way we assure truly reactive balances across the app.

The work:

We were getting a lot of repeated code on providers (aka contexts) so in order to have DRY code, the Network provider was merged into the Wallet provider. The balances mentioned above were also implemented in the Wallet provider.

Wallet provider now exports 4 reactive variables:

- balances:  Balance[] = array of balances for every asset
- connected: boolean = is Marina enabled? (marina.isEnabled() result)
- marina: MarinaProvider = the marina provider (duh)
- network: NetworkString = the current selected network, defaults to testnet

Reactivity works for each variable like this:

- balances: every time Marina emits a NEW_TX event, balances are re-fetched
- connected: every time Marina emits a ENABLED or DISABLED event
- marina: not reactive, will be undefined if user don’t have Marina installed
- network: every time Marina emits a NETWORK event

Other changes worth mentioning:

- New util getAssetBalance(asset, balances), reactive since it uses balances from the Wallet Context
- Renamed wallet variable (from Wallet Context) to connected for clarity
- Pretty function prettyQuantity now accepts a number instead of an Asset
- Removal of functions connect and disconnect from Wallet Context, this functionality is now achieved by listening to ENABLED and DISABLED Marina events
- Renamed util function getBalance(asset) to getAssetBalance(asset, balances) for clarity and reactivity
- Adds defaultNetwork = ‘testnet’ to lib/constants
- Renamed util function getMarina to getMarinaProvider for clarity