Bug fix where contracts were not having fresh info for collateral and synthetic.
In the case of the oracle down, contracts should have collateral value unknown (or 0)
which means the ratio state should be Unknown, and not Safe or Critical as they were.

The problem was contracts were using collateral and synthetic values from the moment
they were created, without updating, which is wrong.